requirements.txt
feedparser
requests
beautifulsoup4
pandas
openai
python-dotenv